Explicit std::min/max template usage
This is unfortunately required in several cases:
 - Sometimes when we use either float and double (this should be changed)
 - Because KDCoordinate is not an int, so any arithmemtic promotes it to
an int
 - Because we mix pointer differences and ints

+++ apps/shared/function_list_controller.cpp
@@ -261,7 +261,7 @@ KDCoordinate FunctionListController::maxFunctionNameWidth() {
     const char * functionName = record.fullName();
     const char * dotPosition = strchr(functionName, Ion::Storage::k_dotChar);
     assert(dotPosition != nullptr);
-    maxNameLength = std::max(maxNameLength, dotPosition-functionName);
+    maxNameLength = std::max(maxNameLength, static_cast<int>(dotPosition-functionName));
   }
   return nameWidth(maxNameLength + Function::k_parenthesedArgumentCodePointLength);
 }
